refactor(tedris): change project structure, group related content under features
