+++ apps/tedris/features/flashcards/components/FlashCard.tsx
