+++ apps/tedris/features/flashcards/components/FlashCardContent.tsx
@@ -1,19 +1,20 @@
 'use client';
 
-import { HadithCard as HadithCardType } from '@madrasah/types';
-import { useState, TouchEvent, MouseEvent } from 'react';
-import { BookOpen, Bookmark } from 'lucide-react';
-import FlashCard from './FlashCard';
-import { useFlashCards } from '../../hooks/useFlashCards';
+import { Bookmark, BookOpen } from 'lucide-react';
+import { MouseEvent, TouchEvent, useState } from 'react';
 
-type HadithCardProps = HadithCardType;
+import { toDisplay } from '../utils/flashCardUtils';
+import { FlashCard } from '@madrasah/types';
 
-export default function HadithCard({ id, fullText, partialText, type }: HadithCardProps) {
+import { useFlashCards } from '../hooks/useFlashCards';
+import FlashCardComponent from './FlashCard';
+
+export default function FlashCardContent(card: FlashCard) {
   const [flipped, setFlipped] = useState(false);
   const [touchStart, setTouchStart] = useState<number | null>(null);
-
+  const data = toDisplay(card);
   const { toggleMemorized, isCardMemorized } = useFlashCards();
-  const memorized = isCardMemorized(id);
+  const memorized = isCardMemorized(data.id);
 
   const handleTouchStart = (e: TouchEvent | MouseEvent) => {
     const clientX = 'touches' in e ? e.touches[0]?.clientX : e.clientX;
@@ -35,7 +36,7 @@ export default function HadithCard({ id, fullText, partialText, type }: HadithCa
   const handleCardFlip = () => setFlipped((prev) => !prev);
 
   return (
-    <div className="relative w-full h-[500px]">
+    <div className="relative h-[500px] w-full">
       <div
         className={`absolute w-full transition-transform duration-500 ease-in-out ${flipped ? 'rotate-y-180' : ''}`}
         style={{
@@ -49,39 +50,45 @@ export default function HadithCard({ id, fullText, partialText, type }: HadithCa
         onMouseLeave={handleTouchEnd}
       >
         {/* Ön Yüz */}
-        <div className="absolute backface-hidden w-full" style={{ backfaceVisibility: 'hidden', WebkitBackfaceVisibility: 'hidden' }}>
-          <FlashCard>
-            <Header />
-            <p className="text-lg sm:text-xl whitespace-pre-wrap break-words text-gray-400">{partialText}</p>
-            <CardActions onFlip={handleCardFlip} memorized={memorized} onToggleMemorized={() => toggleMemorized({ id, type })} />
-          </FlashCard>
+        <div
+          className="backface-hidden absolute w-full"
+          style={{
+            backfaceVisibility: 'hidden',
+            WebkitBackfaceVisibility: 'hidden',
+          }}
+        >
+          <FlashCardComponent>
+            <Header title={data.title} />
+            <p className="whitespace-pre-wrap break-words text-lg text-gray-400 sm:text-xl">{data.question}</p>
+            <CardActions onFlip={handleCardFlip} memorized={memorized} onToggleMemorized={() => toggleMemorized({ id: data.id, type: data.type })} />
+          </FlashCardComponent>
         </div>
 
         {/* Arka Yüz */}
         <div
-          className="absolute rotate-y-180 backface-hidden w-full"
+          className="rotate-y-180 backface-hidden absolute w-full"
           style={{
             backfaceVisibility: 'hidden',
             WebkitBackfaceVisibility: 'hidden',
             transform: 'rotateY(180deg)',
           }}
         >
-          <FlashCard>
-            <Header />
-            <p className="text-lg sm:text-xl whitespace-pre-wrap break-words text-primary font-semibold">{fullText}</p>
-            <CardActions onFlip={handleCardFlip} memorized={memorized} onToggleMemorized={() => toggleMemorized({ id, type })} />
-          </FlashCard>
+          <FlashCardComponent>
+            <Header title={data.title} />
+            <p className="text-primary whitespace-pre-wrap break-words text-lg font-semibold sm:text-xl">{data.value}</p>
+            <CardActions onFlip={handleCardFlip} memorized={memorized} onToggleMemorized={() => toggleMemorized({ id: data.id, type: data.type })} />
+          </FlashCardComponent>
         </div>
       </div>
     </div>
   );
 }
 
-function Header() {
+function Header({ title = '' }) {
   return (
-    <div className="flex items-center justify-center gap-2 text-primary mb-4">
+    <div className="text-primary mb-4 flex items-center justify-center gap-2">
       <BookOpen size={20} />
-      <h3 className="text-base sm:text-lg font-semibold">Hadis Kartı</h3>
+      <h3 className="text-base font-semibold sm:text-lg">{title ?? ''}</h3>
     </div>
   );
 }
@@ -95,17 +102,17 @@ type CardActionsProps = {
 function CardActions({ onFlip, memorized, onToggleMemorized }: CardActionsProps) {
   return (
     <div className="mt-4 flex flex-col items-center gap-4">
-      <button onClick={onFlip} className="p-2 bg-primary text-primary-foreground rounded-full hover:bg-primary/90 transition-colors">
+      <button onClick={onFlip} className="bg-primary text-primary-foreground hover:bg-primary/90 rounded-full p-2 transition-colors">
         <BookOpen size={20} />
       </button>
       <button
         onClick={onToggleMemorized}
-        className={`flex items-center gap-2 px-3 py-2 rounded-full transition-colors ${
+        className={`flex items-center gap-2 rounded-full px-3 py-2 transition-colors ${
           memorized ? 'bg-green-500 text-white hover:bg-green-600' : 'bg-gray-200 text-gray-700 hover:bg-gray-300'
         }`}
       >
         <Bookmark size={16} />
-        <span className="text-sm font-medium">{memorized ? 'Ezberlendi' : 'Ezberle'}</span>
+        <span className="text-sm font-medium">{memorized ? 'Tekrarla' : 'Ezberledim'}</span>
       </button>
     </div>
   );
